requirements.txt
Django==4.2.4
django-filter==23.3
djangorestframework==3.14.0
python-dotenv==1.0.0